Handle .snupkg files
This uses `*.*nupkg` when uploading the artifacts in order to include
the `.snupkg` files (if produced by the build).

The push actions are unchanged; the assumption is that package
repositories with symbol support will trigger a push of the symbols
package as part of the push of the main package.